2019.10.08 변경사항 : 메인 index.jsp -> index2.jsp로 변경, 로그인 AJAX수정 작업 진행중...

+++ Coordination_MVC/src/main/java/com/coordination/main/ClosetController.java
@@ -1,7 +1,11 @@
 package com.coordination.main;
 
+import java.io.File;
 import java.io.PrintWriter;
+import java.util.List;
 
+import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
@@ -10,7 +14,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.servlet.ModelAndView;
 
 import com.coordination.dto.ClosetVO;
 import com.coordination.service.ClosetService;
@@ -18,14 +21,16 @@ import com.coordination.service.ClosetService;
 @Controller
 public class ClosetController {
 
+	@Resource(name="imgPath")
+	private String imgPath;
+	
 	@Autowired
 	private ClosetService service;
 	
 	//회원 - 나만의 옷장 등록
 	@RequestMapping(value = "insertCloset", method = RequestMethod.POST)
-	public ModelAndView insert(ClosetVO vo, HttpServletResponse response, HttpSession session) throws Exception {
-		
-		ModelAndView mav = new ModelAndView();
+	public String insert(ClosetVO vo, Model model, HttpServletResponse response, HttpSession session) throws Exception {
+
 		response.setContentType("text/html; charset=UTF-8");
 		PrintWriter out = response.getWriter();
 		
@@ -57,8 +62,7 @@ public class ClosetController {
             out.flush();
             
             //MyPage로 이동
-            mav.setViewName("movePage");
-            mav.addObject("url", "insertCloset");
+            model.addAttribute("url", "insertCloset");
 			
 		}catch(Exception e) {
 			e.printStackTrace();
@@ -70,12 +74,33 @@ public class ClosetController {
             out.flush();
 		}
 		
-		return mav;
+		return "movePage";
 	}
 	
-	//회원 - 나만의 옷장 수정
+	//회원 - 나만의 옷장 수정1(Form이동)
+	@RequestMapping(value = "updateClosetForm")
+	public String updateForm(ClosetVO vo, Model model, HttpServletRequest request) {
+		
+		int num = 0;
+		num = Integer.parseInt(request.getParameter("num"));
+		
+		try {
+			vo.setNum(num);
+			List<ClosetVO> closetList = service.closetList(vo);
+			
+			model.addAttribute("closetList", closetList);
+			
+		}catch(Exception e) {
+			e.printStackTrace();
+			
+		}
+		
+		return "coordination/member/closetUpdateDelete";
+	}
+	
+	//회원 - 나만의 옷장 수정2
 	@RequestMapping(value = "updateCloset", method = RequestMethod.POST)
-	public String update(ClosetVO vo, HttpServletResponse response, Model model) throws Exception {
+	public String update(ClosetVO vo, Model model, HttpServletResponse response) throws Exception {
 		
 		response.setContentType("text/html; charset=UTF-8");
 		PrintWriter out = response.getWriter();
@@ -105,20 +130,42 @@ public class ClosetController {
 	}
 	
 	//회원 - 나만의 옷장 삭제
-	@RequestMapping(value = "deleteCloset", method = RequestMethod.POST)
-	public String delete(ClosetVO vo, HttpServletResponse response, Model model) throws Exception {
+	@RequestMapping(value = "deleteCloset", method = RequestMethod.GET)
+	public String delete(ClosetVO vo, Model model, HttpServletResponse response, HttpServletRequest request) throws Exception {
+		
+		int num = 0;
+		num = Integer.parseInt(request.getParameter("num"));
+		
+		String img = request.getParameter("img");
+		imgPath += "\\user\\" + img;
 		
 		response.setContentType("text/html; charset=UTF-8");
 		PrintWriter out = response.getWriter();
 		
 		try {
-			service.deleteCloset(vo);
 			
-			out.println("<script>"
-					+ "alert('삭제가 완료되었습니다.');"
-        			+ "</script>");
-            out.flush();
-            
+			//프로젝트 내에 이미지 파일을 삭제하기 위한 파일 객체 선언
+			File file = new File(imgPath);
+			//해당 파일이 존재한다면 DB정보 삭제 + 이미지 삭제
+			if(file.exists() == true)
+			{
+				vo.setNum(num);
+				service.deleteCloset(vo);
+				file.delete();
+				
+				out.println("<script>"
+						+ "alert('삭제가 완료되었습니다.');"
+	        			+ "</script>");
+	            out.flush();
+			}
+			else
+			{
+				out.println("<script>"
+						+ "alert('Error!!!');"
+						+ "history.back();"
+	        			+ "</script>");
+	            out.flush();
+			}
             model.addAttribute("url", "deleteCloset");
 			
 		}catch(Exception e) {
